Fix order of test compare. Make doxygen comments single line. Add MOM prefix to domain files.
